Updated abstract and child classes
